requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
google-generativeai==0.3.2
python-dotenv==1.0.0
reportlab==4.0.7
Werkzeug==3.0.1